compiler error fixed
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@2652 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/tests/cv/src/astereomatching.cpp
@@ -208,7 +208,7 @@ float dispRMS( const Mat& computedDisp, const Mat& groundTruthDisp, const Mat& m
     Mat grayGroundTruthDisp; cvtColor(groundTruthDisp, grayGroundTruthDisp, CV_BGR2GRAY);
 
     int pointsCount = mask.empty() ? computedDisp.cols*computedDisp.rows : countNonZero(mask);
-    return 1.f/sqrt(pointsCount) * norm(grayComputedDisp, grayGroundTruthDisp, NORM_L2, mask);
+    return 1.f/sqrt((float)pointsCount) * norm(grayComputedDisp, grayGroundTruthDisp, NORM_L2, mask);
 }
 
 /*
